[Netbox] Change update trains to 'feature'
Bit of a mistake on my part.
Wanted to try out the new web interface (available in the feature
branch). Of course I forgot to verify that my backup from the stable
train worked, so now I'm temporarily stuck with a 'future' database
version until v3.0 is released and a docker image is published.

I'll be reverting to a more stable release as soon as it's released